+++ packages/coffeescript/package.js
@@ -2,7 +2,7 @@ Package.describe({
   summary: "Javascript dialect with fewer braces and semicolons"
 });
 
-Npm.depends({"coffee-script": "1.5.0"});
+Npm.depends({"coffee-script": "1.6.2"});
 
 var coffeescript_handler = function(bundle, source_path, serve_path, where) {
   var fs = Npm.require('fs');