+++ tools/packages.js
@@ -238,7 +238,13 @@ _.extend(Package.prototype, {
       api.use(project.get_packages(app_dir));
 
       // -- Source files --
-      api.add_files(sources_except(api, "server"), "client");
+      var shouldLoadRaw = function (filename) {
+        return filename.indexOf(path.sep + 'client' + path.sep + 'compatibility' + path.sep) !== -1;
+      };
+      var clientFiles = sources_except(api, "server");
+      api.add_files(_.filter(clientFiles, shouldLoadRaw), "client", {raw: true});
+      api.add_files(_.reject(clientFiles, shouldLoadRaw), "client");
+
       api.add_files(sources_except(api, "client"), "server");
     });